(pid-tools): Persist removing lines (#3254)
Link to JIRA:
https://cognitedata.atlassian.net/browse/CXT-751

# What has changed?

Now if you remove/delete a line in the Diagram Parsing Tool, it will not reappear after the 'Auto analysis' feature is triggered. 

## Why is this important?
The 'Auto analysis' feature often finds more that just the process lines. This can for instance be electrial lines or lines to indicate a note. Now you can remove this unwanted lines and they will not reappear if you want to run 'Auto analysis' again.

# How to test?

1. Open a file and run 'Auto analysis'
2. Delete (at least) one line
3. Run 'Auto analysis' again
4. The expected behaviour now is that the deleted line should not reappaer.
5. (Bonus) Save a parsed graph to CDF
  a. Load the same graph later
  b. Click on 'Auto analysis'
  c. The deleted lines should not reappear