(pid-tools): Improve on hover actions, especially in `addLine` tool (#3256)
This PR belongs together with #3254 to make it more ergonomic to delete lines. They can be review independently on each other.

# Was was changed?

This is mostly a refactor of the `onMouseEnter` function, but is also fixes one bug and adds some new features in the 'addLine' tool:
- Bugfix: The paths are now bolded in the 'addSymbol' tool
- 'addLine' tool: Can add line a new line with shift+mover over
- 'addLine' tool: Can add delete a new line with alt+mover over

From a UX perspective, this is similar to how it's done in the `addSymbol` tool.