(pid-tools): Upsert to DMS from Diagram Parser (#3265)
Link to JIRA:
[CXT-752](https://cognitedata.atlassian.net/browse/CXT-752)

# What has changed?

This PR builds upon #3042 to upsert to DMS directly from the Diagram Parser. The nodes and edges are now uploaded when the user clicks on the "Save" button on the top right corner.

# Any other notes?

I removed the `createHash` function since I couldn't get to work in the browser setting (it appears that is only works in node?).