(pid-tools): `throw` if DMS externalId is over 255 char (#3303)
# What has changed?

DMS doesn't support `externalId`s of over 255 characters, but it does not validate this properly on upserting. Hence, this PR explicitly checks if the `externalId` is over the given limit and `throw` if it's too long.

I also updated the `DMS_SPACE_EXTERNAL_ID` since AFAI understand the current space is in a limbo state since we have uploaded too long IDs. This also requires that we create the new space before this PR is merged.