Pid tools dms beta adaptations (#3346)
# What has changed?
This PR contains fixes for the DMS integration of the pid-tools.
Since yesterday, the FDM Beta was rolled out which included a lot of breaking changes.

Most of the changes in this PR are of permanent nature and will work for a while.

One change which I had to do needs to be revisisted on short notice though: I had to remove the deletion of nodes and edges when (re-)upserting the Graph to DMS. The deletion of nodes with `externalId` of length > 64 chars is currently broken. A fix is underway as described in this slack thread https://cognitedata.slack.com/archives/C03G11UNHBJ/p1663678496594659
Once it is rolled out, we can activate the deletion again.