More parellalization in graph upsert and measurement of timings (#3355)
# What has changed?

Some tweaks to make the upserting of the graph to DMS somewhat faster.
- Replacing simple for loops with `awaits` with `Promise.all`
- Basically we now  parallelize all delete and upsert requests for different node or edge types. 
- As long as the server is able to handle those parallel requests, this should lead to a speedup.
Previously, I had removed this parallelization as it led to 409 errors. Those issues seem to be fixed upstream, so it seems save to activate it again.

I also increased the verbosity by printing timings of the upsert to the console. I think it is interesting to see how fast it is and whether the timings change, so that we can raise if it becomes slower.