fix(UFV): fix delete and cancel button (#488)
* fix: use ant modal instead of cogs.js

Using the cogs.js modal together with the antd modal resulted in the
modal not showing up in cdf-ui-interactive-diagrams. However, using the
antd modal instead worked fine.

I aøso tried using the cogs.js dropdown instead, but that didn't work
either in cdf-ui-interactive-diagrams (the modal didnt show up).

* fix: deselect pending annotation on cancel

The cancel button wasn't working correctly in the UFV file preview since
we didn't deselect the file. Moreover, when an annotation for the file
is succesfully deleted the modal is now properly closed.

* fix: deselect annotations on mode change + file change

Co-authored-by: Marvin Reza <marvinreza@Marvins-MacBook-Pro.local>
Co-authored-by: cognite-bulldozer[bot] <51074376+cognite-bulldozer[bot]@users.noreply.github.com>