feat: [UFV-203] Use UFV as primary file viewer (#606)
This PR replaces the old file viewer (using react-picture-annotation) with UFV. This is achieved by bumping the version of `@cognite/data-exploration` and replacing all instances of `*Preview` with `*PreviewUFV` (the UFV-prefix will be removed in a later stage once testing and validation is completed). Finally, since the previous version of this app used rather old versions of cogs.js and data-exploration-components, the style of some components were broken. We have, accordingly, added fixes to these components where it was deemed necessary.