fix: Make badge more independent of scale (#539)
Ideally, we would like the badge to be fully independent of the UFV
zoom level. However, UFV does not support this so this is the second
best solution. This PR adds a larger border with the same color as the badge.
Since the border is independent of the zoom level, the badge will
appear larger when zoomed out and smaller when zoomed in.

The badge radius is also decreased — this makes the badge appear
smaller when zoomed in.

Co-authored-by: cognite-bulldozer[bot] <51074376+cognite-bulldozer[bot]@users.noreply.github.com>