fix(UFV): re-render only when page or file changes (#541)
Previously, we would re-render containers every 30s since CDF file
urls expire after 30s. If we did not do this, page switching would for
example fail (since we need to get the page from the url). Another issue
was that the container and its annotations would blink every 30s due to
the re-render that would need to happen.

This work fixes the mentioned issue by using the new UFV utility
getContainerConfigFromFileInfo. The utility generates a new file url
only when necessary (i.e., every 30s for CDF files) and is meant to be
used *on-demand. Consequently, the file urls are only refreshed when we
actually need to re-render the container (e.g., on file/page change).