Various cognuit fixes (#3658)
* Add Api endpoint for regenerating a single object translation

This is for future use.

* Don't use pagination on configuration page)

* Don't use pagination for data transfers to make it easier for the user to get an overview

This should really use server-side pagination, but as a temporary fix,
this works.

* Remove the reference to Discover

It was meant to be a panel in Discover for Aker BP, but that
idea never materialized. Keeping the logo use, though, as we
don't have a custom logo. :-P

* Remove not implemented Openworks to Petrel Studio option

Due to the trials and tribulations of Halliburton, we never
had time to implement the sync back from Openworks to
Petrel Studio.

So disabling it for now. Maybe(tm) one time in the future.. (or
a different sync).

* Remove unused (and thus confusing) configuration for filtering

This was probably meant to centralize sorting configs in one place,
but it was never used. Removing, since it was very confusion (also
because one of the columns, 'author', should have been
'source.author')

* Allow it to be configurable parent in column header

Make it possbible to remove the parent annotation in the column
header.

In this case, since we only have one source and only talk about Petrel
Studio to Openworks, source is by default the correct interpreation,
and it only gets in the way of sorting/filtering icons.

* Make it possible to filter on datatype (common usage)

* Arrays of keys are required by @tanstack/query v4 or newer (not strings)

This was fixed for everything EXCEPT data statuses, which meant we
didn't get any data statuses back.

* Refresh configurations and data transfers automatically

We've made /configurations and /datatransfers a lot more efficient,
and it is really annoying to have to reload the whole app to force a
refresh.

This refreshes configuration every 20s second, and /datatransfers (the
logs) every 30s. We might have to reduce the frequency later..