fix(platypus): disable close button while publishing a new data model version [UX-41] (#1199)
* Remove dead data set code from data model modal

* User can't close DM create modal while loading

* Don't mount CreateDataModel unless we render it

If we keep it rendered but only toggle visibility, then anything the user enters will persist after the modal is closed and reopened. Instead we mount and unmount the modal when toggling its visibility. This was the behavior we had prior to the Cogs v8 PR (https://github.com/cognitedata/fusion/pull/869/)

---------

Co-authored-by: 23andreas <andreas.skonberg@gmail.com>