Support rendering lists of time series.
This renders lists of timeseries in the preview, prefixed by
external-id. The code isn't particularly nice, I must admit, but it
works instead of showing a JSON list of external-ids like before.

The worry is that we special case 'timeseries' a lot of places (and
with various casing based on where it comes from), and this means we
will need to add the same special casing for other resource types.

We should have a similar preview for files (showing a rendering of the
first page?).