Handle lists of files and sequences in data management tab
This is a temporary fix to avoid errors when double clicking on fields
representing a file, list of files, a sequence and list of sequences,
that are referenced.

The panel that appears is not particularly interesting, but it is
better than causing the frontend to crash. Also, it is a placeholder
for integrating a better preview in a future commit.