Handle non-resolved CDF external-id references
There is no guarantee for referencial integrity for the external-id,
so the reference is then returned as null from the mixer. We did want
to filter it out server-side, but that isn't in line with how GraphQL
is expected to work (the number of inputs needs to match the output).

Instead we return a descriptive text. We cannot include the
external-id that failed, as we don't have it (the schema service
doesn't support returning it when using the query interface).