fix Autocomplete and Select menu positioning
I set menuPortalTarget for Select so that the menu is opened in a portal and the absolute positioning style is applied correctly. I saw it was done this way in the platypus app in the fusion monorepo.

I also refactored some AutoComplete usages to use Select instead since AutoComplete is deprecated. I didn't do that for TimeSeriesSelector because of time constraints but it could be done there as well.