Tidy up handling of view versions (#2831)
Turns out the view version is a different thing from the data model
version. Not sure if it was always like this, or if this is a recent
change.

This commit changes a lot of code that tries to pick out view properties
from the url, to use a new hook called useCurrentView. This gives
the current view in most situations, and the current view usually
contains all the properties we need (view external id, view version and
space).

Jira: https://cognitedata.atlassian.net/browse/CXT-1309

Co-authored-by: cognite-bulldozer[bot] <51074376+cognite-bulldozer[bot]@users.noreply.github.com>