Add the useCurrentLinkedView hook
In a previous commit, I introduced the useCurrentView hook, but was a
bit heavy-handed in introducing it, and ended up using the current view
also in situations where I should have been referring to the linked view
(the one referred to by the currently selected direct relation
property).

The useCurrentLinkedView hook behaves like the useCurrentView hook, but
gives you the linked view instead.