chore(cog-idp): Highlighting the Cog-IdP login button (#3110)
* chore(cog-idp): Refactoring to prepare for Cog-IdP login

This PR does not change any behavior or UI, but extracts the rendering of the login buttons to a separate function, to declutter the upcoming PR where we will actually change the UI for organizations where login via the Cognite Identity Provider has been enabled.

Issue: https://cognitedata.atlassian.net/browse/AUT-335

* Fix lint

* chore(cog-idp): Highlighting the Cog-IdP login button

For organizations (domains) that we consider to be "internal" and where DLC says that there is a default IdP (which will in practice always be Cog-IdP), we will show only that IdP button and hide the others in an accordion that is initially closed. This will be used during the migration period where we gradually roll out Cog-IdP, so that users can fall back to the old login if something is wrong with the new one.

This is based on the preliminary refactor in https://github.com/cognitedata/fusion/pull/3108.

Issue: https://cognitedata.atlassian.net/browse/AUT-335

* Fix lint

* Check explicitly for Cog-IdP

* feat: check if org exist in CogIdP

* chore: review fixes

* chore: review fixes 2

---------

Co-authored-by: NilsBarlaug <nils.barlaug@cognite.com>
Co-authored-by: NilsBarlaug <NilsBarlaug@users.noreply.github.com>
Co-authored-by: cognite-bulldozer[bot] <51074376+cognite-bulldozer[bot]@users.noreply.github.com>