fix(entity-matching): Set default sort orders (#3309)
Most of the tables in entity matching are better if they sort by
confidence or by number of matches descending, so that the most relevant
entries show up first by default. This is how the old entity matching
app works and it's something that's been annoying me with the new app.

Co-authored-by: cognite-bulldozer[bot] <51074376+cognite-bulldozer[bot]@users.noreply.github.com>