fix(interactive-diagrams): reduce excessive API calls (#3845)
* fix(interactive-diagrams): reduce excessive API calls

The Interactive Diagrams app landing page makes a single call to the events
(and annotations) API for every single pending diagram in the world,
regardless of whether or not the diagram is actually shown on the page.

The main reason for this seems to have been to support sorting on number
of annotations.

This commit removes the excessive calls and also removes the ability to
sort on number of annotations on a pending diagram in the list.

* fix e2e

---------

Co-authored-by: cognite-bulldozer[bot] <51074376+cognite-bulldozer[bot]@users.noreply.github.com>